Added support for basic XQuery FLOWR expressions; new serializer for serializing the DOM tree back into XML; many bug-fixes.
svn path=/trunk/eXist-1.0/; revision=133


Former-commit-id: ad9d55f515de2cb1fa50cf09c9e42150efc8374f

+++ src/org/exist/xpath/UnaryExpr.java
@@ -23,7 +23,7 @@
 package org.exist.xpath;
 
 import org.exist.dom.DocumentSet;
-import org.exist.xpath.value.DecimalValue;
+import org.exist.xpath.value.DoubleValue;
 import org.exist.xpath.value.Item;
 import org.exist.xpath.value.Sequence;
 import org.exist.xpath.value.Type;
@@ -47,7 +47,7 @@ public class UnaryExpr extends PathExpr {
 			contextSequence = contextItem.toSequence();
 		if(getLength() == 0)
 			throw new XPathException("unary expression requires an operand");
-		DecimalValue value = (DecimalValue)
+		DoubleValue value = (DoubleValue)
 			getExpression(0).eval(context, docs, contextSequence).convertTo(Type.DECIMAL);
 		switch(mode) {
 			case Constants.MINUS :
